[sw] Remove hard coded UART dependency
We should be able to build UART-less tops for FPGA and simulation
devices.
