+++ sw/device/lib/arch/device_fpga_cw310.c
@@ -9,8 +9,6 @@
 #include "sw/device/silicon_creator/lib/drivers/ibex.h"
 #include "sw/device/silicon_creator/lib/drivers/uart.h"
 
-#include "uart_regs.h"
-
 /**
  * @file
  * @brief Device-specific symbol definitions for the ChipWhisperer CW310 device.
@@ -50,24 +48,9 @@ const uint32_t kUartBaud1M33 =
 const uint32_t kUartBaud1M50 =
     CALCULATE_UART_NCO(1500000, kClockFreqPeripheralHz);
 
-const uint32_t kUartTxFifoCpuCycles = CALCULATE_UART_TX_FIFO_CPU_CYCLES(
-    kUartBaudrate, kClockFreqCpuHz, UART_PARAM_TX_FIFO_DEPTH);
-
 const uint32_t kAstCheckPollCpuCycles =
     CALCULATE_AST_CHECK_POLL_CPU_CYCLES(kClockFreqCpuHz);
 
 uintptr_t device_test_status_address(void) { return 0; }
 
 uintptr_t device_log_bypass_uart_address(void) { return 0; }
-
-void device_fpga_version_print(void) {
-  // This value is guaranteed to be zero on all non-FPGA implementations.
-  uint32_t fpga = ibex_fpga_version();
-  //                       : M O R
-  const uint32_t kRom = 0x3a4d4f52;
-  uart_write_imm(kRom);
-  // The cast to unsigned int stops GCC from complaining about uint32_t
-  // being a `long unsigned int` while the %x specifier takes `unsigned int`.
-  const uint32_t kNewline = 0x0a0d;
-  uart_write_hex(fpga, sizeof(fpga), kNewline);
-}
