+++ sw/device/lib/arch/device_silicon.c
@@ -6,9 +6,6 @@
 
 #include "sw/device/lib/arch/device.h"
 
-#include "rv_core_ibex_regs.h"
-#include "uart_regs.h"
-
 /**
  * Device-specific symbol definitions for the Silicon device.
  */
@@ -47,14 +44,9 @@ const uint32_t kUartBaud1M33 =
 const uint32_t kUartBaud1M50 =
     CALCULATE_UART_NCO(1500000, kClockFreqPeripheralHz);
 
-const uint32_t kUartTxFifoCpuCycles = CALCULATE_UART_TX_FIFO_CPU_CYCLES(
-    kUartBaudrate, kClockFreqCpuHz, UART_PARAM_TX_FIFO_DEPTH);
-
 const uint32_t kAstCheckPollCpuCycles =
     CALCULATE_AST_CHECK_POLL_CPU_CYCLES(kClockFreqCpuHz);
 
 uintptr_t device_test_status_address(void) { return 0; }
 
 uintptr_t device_log_bypass_uart_address(void) { return 0; }
-
-void device_fpga_version_print(void) {}
