+++ sw/device/lib/arch/device_sim_dv.c
@@ -8,7 +8,6 @@
 #include "sw/device/lib/arch/device.h"
 
 #include "rv_core_ibex_regs.h"
-#include "uart_regs.h"
 
 // Use the first dt_rv_core_ibex_t enum, i.e. the first Ibex core instance.
 static const dt_rv_core_ibex_t kRvCoreIbexDt = (dt_rv_core_ibex_t)0;
@@ -59,9 +58,6 @@ const uint32_t kUartBaud1M33 =
 const uint32_t kUartBaud1M50 =
     CALCULATE_UART_NCO(1500000, kClockFreqPeripheralHz);
 
-const uint32_t kUartTxFifoCpuCycles = CALCULATE_UART_TX_FIFO_CPU_CYCLES(
-    kUartBaudrate, kClockFreqCpuHz, UART_PARAM_TX_FIFO_DEPTH);
-
 const uint32_t kAstCheckPollCpuCycles =
     CALCULATE_AST_CHECK_POLL_CPU_CYCLES(kClockFreqCpuHz);
 
@@ -72,5 +68,3 @@ uintptr_t device_test_status_address(void) {
 uintptr_t device_log_bypass_uart_address(void) {
   return rv_core_ibex_base() + RV_CORE_IBEX_DV_SIM_WINDOW_REG_OFFSET + 0x04;
 }
-
-void device_fpga_version_print(void) {}
